refactor abi encoding trait
