+++ src/codegen/events/solana.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::codegen::cfg::{ControlFlowGraph, Instr};
-use crate::codegen::encoding::create_encoder;
+use crate::codegen::encoding::abi_encode;
 use crate::codegen::events::EventEmitter;
 use crate::codegen::expression::expression;
 use crate::codegen::vartable::Vartable;
@@ -46,9 +46,8 @@ impl EventEmitter for SolanaEventEmitter<'_> {
         let mut to_be_encoded: Vec<Expression> = vec![discriminator];
         to_be_encoded.append(&mut codegen_args);
 
-        let mut encoder = create_encoder(self.ns, false);
         let (abi_encoded, abi_encoded_size) =
-            encoder.abi_encode(&self.loc, to_be_encoded, self.ns, vartab, cfg);
+            abi_encode(&self.loc, to_be_encoded, self.ns, vartab, cfg, false);
 
         cfg.add(
             vartab,
