+++ src/codegen/expression.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
+use super::encoding::abi_encode;
 use super::storage::{
     array_offset, array_pop, array_push, storage_slots_array_pop, storage_slots_array_push,
 };
@@ -666,16 +667,16 @@ pub fn expression(
             payable_transfer(args, cfg, contract_no, func, ns, vartab, loc, opt)
         }
         ast::Expression::Builtin(loc, _, ast::Builtin::AbiEncode, args) => {
-            abi_encode(args, cfg, contract_no, func, ns, vartab, loc, opt)
+            abi_encode_many(args, cfg, contract_no, func, ns, vartab, loc, opt)
         }
         ast::Expression::Builtin(loc, _, ast::Builtin::AbiEncodePacked, args) => {
-            abi_encode_packed(args, cfg, contract_no, func, ns, vartab, loc, opt)
+            abi_encode_packed_expression(args, cfg, contract_no, func, ns, vartab, loc, opt)
         }
         ast::Expression::Builtin(loc, _, ast::Builtin::AbiEncodeWithSelector, args) => {
-            abi_encode_with_selector(args, cfg, contract_no, func, ns, vartab, loc, opt)
+            abi_encode_expression_with_selector(args, cfg, contract_no, func, ns, vartab, loc, opt)
         }
         ast::Expression::Builtin(loc, _, ast::Builtin::AbiEncodeWithSignature, args) => {
-            abi_encode_with_signature(args, loc, cfg, contract_no, func, ns, vartab, opt)
+            abi_encode_expression_with_signature(args, loc, cfg, contract_no, func, ns, vartab, opt)
         }
         ast::Expression::Builtin(loc, _, ast::Builtin::AbiEncodeCall, args) => {
             abi_encode_call(args, cfg, contract_no, func, ns, vartab, loc, opt)
@@ -1266,7 +1267,7 @@ fn payable_transfer(
     Expression::Poison
 }
 
-fn abi_encode(
+fn abi_encode_many(
     args: &[ast::Expression],
     cfg: &mut ControlFlowGraph,
     contract_no: usize,
@@ -1281,11 +1282,10 @@ fn abi_encode(
         .map(|v| expression(v, cfg, contract_no, func, ns, vartab, opt))
         .collect::<Vec<Expression>>();
 
-    let mut encoder = create_encoder(ns, false);
-    encoder.abi_encode(loc, args, ns, vartab, cfg).0
+    abi_encode(loc, args, ns, vartab, cfg, false).0
 }
 
-fn abi_encode_packed(
+fn abi_encode_packed_expression(
     args: &[ast::Expression],
     cfg: &mut ControlFlowGraph,
     contract_no: usize,
@@ -1300,12 +1300,11 @@ fn abi_encode_packed(
         .map(|v| expression(v, cfg, contract_no, func, ns, vartab, opt))
         .collect::<Vec<Expression>>();
 
-    let mut encoder = create_encoder(ns, true);
-    let (encoded, _) = encoder.abi_encode(loc, packed, ns, vartab, cfg);
+    let (encoded, _) = abi_encode(loc, packed, ns, vartab, cfg, true);
     encoded
 }
 
-fn encode_with_selector(
+fn encode_many_with_selector(
     loc: &pt::Loc,
     selector: Expression,
     mut args: Vec<Expression>,
@@ -1316,11 +1315,10 @@ fn encode_with_selector(
     let mut encoder_args: Vec<Expression> = Vec::with_capacity(args.len() + 1);
     encoder_args.push(selector);
     encoder_args.append(&mut args);
-    let mut encoder = create_encoder(ns, false);
-    encoder.abi_encode(loc, encoder_args, ns, vartab, cfg).0
+    abi_encode(loc, encoder_args, ns, vartab, cfg, false).0
 }
 
-fn abi_encode_with_selector(
+fn abi_encode_expression_with_selector(
     args: &[ast::Expression],
     cfg: &mut ControlFlowGraph,
     contract_no: usize,
@@ -1343,10 +1341,10 @@ fn abi_encode_with_selector(
     let args = args_iter
         .map(|v| expression(v, cfg, contract_no, func, ns, vartab, opt))
         .collect::<Vec<Expression>>();
-    encode_with_selector(loc, selector, args, ns, vartab, cfg)
+    encode_many_with_selector(loc, selector, args, ns, vartab, cfg)
 }
 
-fn abi_encode_with_signature(
+fn abi_encode_expression_with_signature(
     args: &[ast::Expression],
     loc: &pt::Loc,
     cfg: &mut ControlFlowGraph,
@@ -1368,7 +1366,7 @@ fn abi_encode_with_signature(
     let args = args_iter
         .map(|v| expression(v, cfg, contract_no, func, ns, vartab, opt))
         .collect::<Vec<Expression>>();
-    encode_with_selector(loc, selector, args, ns, vartab, cfg)
+    encode_many_with_selector(loc, selector, args, ns, vartab, cfg)
 }
 
 fn abi_encode_call(
@@ -1399,7 +1397,7 @@ fn abi_encode_call(
     let args = args_iter
         .map(|v| expression(v, cfg, contract_no, func, ns, vartab, opt))
         .collect::<Vec<Expression>>();
-    encode_with_selector(loc, selector, args, ns, vartab, cfg)
+    encode_many_with_selector(loc, selector, args, ns, vartab, cfg)
 }
 
 fn builtin_evm_gasprice(
@@ -2201,8 +2199,7 @@ pub fn emit_function_call(
                     args,
                 ];
 
-                let mut encoder = create_encoder(ns, true);
-                let (encoded, _) = encoder.abi_encode(loc, encoder_args, ns, vartab, cfg);
+                let (encoded, _) = abi_encode(loc, encoder_args, ns, vartab, cfg, true);
                 (encoded, None)
             } else {
                 (args, Some(address))
@@ -2309,8 +2306,7 @@ pub fn emit_function_call(
                     Some(address)
                 };
 
-                let mut encoder = create_encoder(ns, false);
-                let (payload, _) = encoder.abi_encode(loc, args, ns, vartab, cfg);
+                let (payload, _) = abi_encode(loc, args, ns, vartab, cfg, false);
 
                 cfg.add(
                     vartab,
@@ -2401,8 +2397,7 @@ pub fn emit_function_call(
                     Some(address)
                 };
 
-                let mut encoder = create_encoder(ns, false);
-                let (payload, _) = encoder.abi_encode(loc, args, ns, vartab, cfg);
+                let (payload, _) = abi_encode(loc, args, ns, vartab, cfg, false);
 
                 cfg.add(
                     vartab,
@@ -2910,8 +2905,7 @@ pub(super) fn assert_failure(
     let selector = Expression::NumberLiteral(Loc::Codegen, Type::Uint(32), BigInt::from(selector));
     let args = vec![selector, arg.unwrap()];
 
-    let mut encoder = create_encoder(ns, false);
-    let (encoded_buffer, _) = encoder.abi_encode(loc, args, ns, vartab, cfg);
+    let (encoded_buffer, _) = abi_encode(loc, args, ns, vartab, cfg, false);
 
     cfg.add(
         vartab,
